Adobe-India-Hackathon25/Challenge_1b/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
deepsearch-glm==0.26.1
docling==2.4.2
docling-core==2.3.1
docling-ibm-models==2.0.3
docling-parse==2.0.3
docutils==0.21.2
easyocr==1.7.2
einops==0.8.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-huggingface==0.1.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.142
lazy_loader==0.4
lxml==4.9.4
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mean-average-precision==2021.4.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbformat==5.10.4
networkx==3.4.2
ninja==1.11.1.1
numpy>=1.21.0
# GPU dependencies removed for CPU-only operation
ollama==0.3.3
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
pyarrow==16.1.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pypdfium2==4.30.0
PyPDF2==3.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
# triton==3.1.0  # Removed for CPU-only operation - GPU dependency
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
XlsxWriter==3.2.0
yarl==1.17.1
streamlit==1.40.0